+++ categories/urls.py
@@ -1,7 +1,6 @@
 from django.urls import path
 
-from .views      import ProductListView, SubCategoriesView
+from .views      import SubCategoriesView
 urlpatterns = {
-  path('/product', ProductListView.as_view()),
   path('/subcategory', SubCategoriesView.as_view()),
 }
