+++ julabo/urls.py
@@ -3,5 +3,5 @@ from django.urls import path,include
 urlpatterns = [
   path('categories', include('categories.urls')),
   path('users', include('users.urls')),
-  path('productgroups', include('products.urls')),
+  path('products', include('products.urls')),
 ]
