+++ products/views.py
@@ -28,3 +28,36 @@ class ProductGroupDetailView(View):
         mls    = [result['ml'] for result in results]
         prices = [result['price'] for result in results]
         return JsonResponse({'products': results, 'mls': mls, 'prices': prices}, status=200)
+
+
+class ProductListView(View):
+  def get(self,request):
+    try:
+      category_subcategory_id = request.GET.get('category_subcategory_id', None)
+      type_ml                 = request.GET.get('ml', None)
+      image                   = Image.objects.all()
+      
+      filter_set = {
+        'categorysubcategory_id': category_subcategory_id,
+        'ml'                    : type_ml
+      }
+      
+      products = Product.objects.filter(**filter_set) 
+
+      products = [{   
+        'id'             : product.id,
+        'name'           : product.name,
+        'ml'             : product.ml,
+        'price'          : product.price,
+        'productgroup_id': product.productgroup.id,
+        'image'          : [img for img in image.filter(product_id = product.id).values('image_url')],
+        'subcategory': {
+          'subcategory_id'  : product.categorysubcategory.subcategory.id,
+          'subcategory_name': product.categorysubcategory.subcategory.name
+        }
+      } for product in products]
+
+      return JsonResponse({'products': products}, status= 200)
+
+    except:
+      return JsonResponse({'message': 'KEY_ERROR'}, status=400)
